requirements.txt
ml-collections
absl-py
diffusers[torch]==0.17.1
accelerate==0.17
wandb
torchvision
inflect==6.0.4
pydantic==1.10.9
transformers==4.30.2
ipdb
albumentations